test_install_errors.xul

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<window title="Mozilla Bug "
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"/>
  <script type="application/javascript" src="apphelper.js"/>
  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id="
     target="_blank">Mozilla Bug </a>
  </body>

<script type="text/javascript"> 
<![CDATA[

SimpleTest.waitForExplicitFinish();

try { 
  navigator.mozApps.install();
} catch (e) {
  ok(e.message == "Not enough arguments \[mozIDOMApplicationRegistry.install\]", "install returned " + e.message);
}

mozAppscb(navigator.mozApps.install(SERVERS['json_syntax_error'], null), 
          { status: "== \"error\"",
            name: "== \"MANIFEST_PARSE_ERROR\""});

mozAppscb(navigator.mozApps.install(SERVERS['missing_required_field'], null),
          { status: "== \"error\"",
            name: "== \"INVALID_MANIFEST\""});

mozAppscb(navigator.mozApps.install(SERVERS['no_delegated_install'], null),
          { status: "== \"error\"",
            name: "== \"PERMISSION_DENIED\""});

clickPopup();
mozAppscb(navigator.mozApps.install(SERVERS['bad_content_type'], null),
          { status: "== \"error\"",
            name: "== \"INVALID_MANIFEST\""},tearDown);



function tearDown() {
  info("in " + arguments.callee.name);
  uninstallAll(function() {SimpleTest.finish();});
}


]]>

</script> 

</window>
